App-specific LLM setting is in dialog.
