+++ src/appMetadata/fetchAppMetaData.ts
@@ -4,7 +4,7 @@ import { getBaseUrl } from "@/components/decentBar/decentBarUtil";
 /* v8 ignore start */
 export async function fetchAppMetadataText():Promise<string> {
   const baseUrl = getBaseUrl();
-  const appMetaDataUrl = `${baseUrl}app-metadata.json`;
+  const appMetaDataUrl = `${baseUrl}/app-metadata.json`;
   const response = await fetch(appMetaDataUrl);
   if (!response.ok) throw new Error(`Failed to fetch app metadata from ${appMetaDataUrl}: ${response.status} ${response.statusText}`);
   return await response.text();
